% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_data.R
\name{find_symbol_id}
\alias{find_symbol_id}
\title{Find the Internal ID of a Symbol}
\usage{
find_symbol_id(symbol)
}
\arguments{
\item{symbol}{A character string representing the symbol of the stock.}
}
\description{
Find the Internal ID of a Symbol
}
