% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_data.R
\name{get_historical_data}
\alias{get_historical_data}
\title{Get Historical Data}
\usage{
get_historical_data(
  symbol,
  start_date = Sys.Date() - lubridate::years(3),
  end_date = Sys.Date()
)
}
\arguments{
\item{symbol}{A character string representing the symbol of the stock.}

\item{start_date}{A character string representing the start date of the data.}

\item{end_date}{A character string representing the end date of the data.}
}
\value{
A data frame containing the historical data of the stock.
}
\description{
Retrieves historical data of a stock from the Casablanca Stock Exchange.
}
\note{
The format of start_date and end_date should be "YYYY-MM-DD".

Use the function get_symbols() to get the list of available symbols.

The API returns only the previous 3 years of data.
}
