% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols.R
\name{get_symbols}
\alias{get_symbols}
\title{Get Available Symbols From Casablanca Stock Exchange}
\usage{
get_symbols()
}
\value{
A tibble with the following columns: name, id, isin, ipo_date,
shares_count, ipo_price, nominal_value, shares_number.
}
\description{
This function retrieves all traded symbols in Casablanca Stock Exchange.
}
