% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_data.R
\name{parse_date}
\alias{parse_date}
\title{Parse Date}
\usage{
parse_date(date)
}
\arguments{
\item{date}{A character string representing the date.}
}
\value{
A date object.
}
\description{
Parse start_date and end_date in the correct format.
}
