% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_data.R
\name{url_constructor}
\alias{url_constructor}
\title{Construct URL for Historical Data}
\usage{
url_constructor(symbol, start_date, end_date)
}
\arguments{
\item{symbol}{A character string representing the symbol of the stock.}

\item{start_date}{A character string representing the start date of the data.}

\item{end_date}{A character string representing the end date of the data.}
}
\description{
Construct URL for Historical Data
}
\note{
The format of start_date and end_date should be "YYYY-MM-DD".
}
