% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_data.R
\name{validate_dates}
\alias{validate_dates}
\title{Validate Dates}
\usage{
validate_dates(start_date, end_date)
}
\arguments{
\item{start_date}{A date object representing the start date.}

\item{end_date}{A date object representing the end date.}
}
\value{
TRUE if the dates are valid.
}
\description{
Check if start_date and end_date are valid.
}
